Move LauncherUtils to package specific to utils
This also fixes the need to add new dependency in all custom schedulers.
