+++ heron/schedulers/src/java/com/twitter/heron/scheduler/slurm/SlurmLauncher.java
@@ -18,7 +18,7 @@ import java.nio.file.Paths;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
-import com.twitter.heron.scheduler.LauncherUtils;
+import com.twitter.heron.spi.utils.LauncherUtils;
 import com.twitter.heron.spi.common.Config;
 import com.twitter.heron.spi.common.Context;
 import com.twitter.heron.spi.packing.PackingPlan;
